Ifpack2: Added more macros for pkg deps to auto-gen'd config header.
Ifpack2 currently has an explicit required dependency on
KokkosClassic.  I would like to remove that in preparation for
deprecating KokkosClassic.  One way to start is to protect all use of
KokkosClassic in Ifpack2 with a macro, but that macro doesn't exist
yet.  TriBits helps generate those macros, but they have to be added
explicitly to the (sub)package's _config.h.in file in order to exist.
This commit adds several macros to that file, and also reorganizes it
and adds comments.