KokkosClassic: Removed TSQR-specific CMake config for BLAS & LAPACK.
TSQR used to have a lot of CMake checks for the existence of specific
LAPACK routines.  This is because TSQR prefers QR factorizations that
produce an R factor with a nonnegative diagonal.  That generated a lot
of superfluous output at Trilinos configure time, for no good reason.
This commit removes all of those tests, and also removes TSQR's own
BLAS and LAPACK wrappers.  TSQR now relies on Teuchos' BLAS and LAPACK
wrappers.

This commit also removes CMake support for the Fortran 9x version of
TSQR's combine kernels (wrapped by TSQR::CombineFortran).  If we want
to generalize TSQR to support arbitrary array layouts (e.g., both
row-major and column-major), we'll have to use Kokkos::View in C++.
I've left the Fortran 9x kernels in place in case we want to
experiment with them later.

NOTE: If we want to support arbitrary array layouts, we may need
limited BLAS and LAPACK support for that.  This may imply hand-rolled
kernels.

NOTE: For computing an R factor with a nonnegative diagonal, we prefer
the run-time approach of testing whether the diagonal satisfies the
requirement, and if not, scaling it.  This can be done on each MPI
process independently, assuming that the signs are the same on all
processes.  If the signs are _not_ the same, it's likely that the
corresponding diagonal entries are close to zero anyway; we defer
considering the numerical analysis implications.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Anasazi
Disabled Packages: FEI,Moertel,MOOCHO,NOX,Rythmos,STK
0) MPI_DEBUG => passed: passed=224,notpassed=0 (14.58 min)
1) SERIAL_RELEASE => passed: passed=205,notpassed=0 (13.11 min)
Other local commits for this build/test group: d051894, 108e399, ef25302, 89cd24a