[Sacado,Stokos] hotfix for missing symbols in linking
This fixes some missing symbols at link time for Sacado and Stokhos. Especially
for the Stokhos case, it's not clear to me why the library isn't linked in
automatically, so probably this hotfix will have to revised in the future.
Anyhow, it doesn't hurt to add an extra DEPLIB for now.

Build/Test Cases Summary
Enabled Packages: Sacado, Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=502,notpassed=0 (2.91 min)
1) SERIAL_RELEASE => passed: passed=504,notpassed=0 (1.38 min)