Kokkos: started to add sorting algorithms
Kokkos_Sort.hpp (part of kokkos/algorithms) will provide sorting interfaces for Kokkos views.
The initial commit provides an interface to std::sort on the host for 1D Views and
a 1D and 3D spatial binning sort nativly implemented in Kokkos. The latter two can be used to
create permutation vectors which can be used to sort multiple arrays the same way. A possible
usecase is particle data. First the permutation vector is generated from the positions using
a binning algorithm, then the permutation vector is used to sort all particle arrays. For
example in 3D:
    // How many bins should be used in each dimensions
    int bin_max[3] = {bin_1d,bin_1d,bin_1d};
    // Minimum values for the three dimensions of the key array (the particle positions)
    double min[3] = {0,0,0};
    // Maxmimum values
    double max[3] = {100.0,100.0,100.0};
    // The key array type (positions of particles)
    typedef Kokkos::View<double*[3]> PosArray;
    // The binning operand, it is used to create the bin index from the positions
    typedef Kokkos::SortImpl::DefaultBinOp3D<PosArray > BinOp;
    BinOp bin_op(bin_max,min,max);
    // The sorting class - no sort within the bins required
    Kokkos::BinSort< PosArray , BinOp >
      Sorter(pos,bin_op,false);
    // Create Permutation Vector
    Sorter.create_permute_vector();
    // Sort position array
    Sorter.sort<PosArray >(pos);
    // Sort velocity array
    Sorter.sort<PosArray >(velocity);

BinSort is templated on the BinningOperation and the Key type.
It has a number of functions for the user:
  // Create the permutation vector, the bin_offset array and the bin_count array. Can be called again if keys changed
  void create_permute_vector();

  // Sort a view with respect ot the first dimension using the permutation array
  template<class ValuesViewType>
  void sort(ValuesViewType values);

  // Get the permutation vector
  KOKKOS_INLINE_FUNCTION
  offset_type get_permute_vector() const;

  // Get the start offsets for each bin
  KOKKOS_INLINE_FUNCTION
  offset_type get_bin_offsets() const;

  // Get the count for each bin
  KOKKOS_INLINE_FUNCTION
  bin_count_type get_bin_count() const;

There is also a function:
template<class 1DViewType>
Kokkos::sort(1DViewType array, bool force_kokkos=false);

If force_kokkos==true or the view is not accessible from the host it uses a 1D Binned sort
to sort the data. Otherwise it uses std::sort.
The binned sorting assumes relativly homogenous distribution. The number of bins is chosen
as number of elements/2. Note that this function works only if the scalar type of the view
is a number (i.e. comparison, addtion, division, multiplication and conversion to double are
defined).