Added simple unit-tests to compute saxpy in both Kokkos and raw Cuda.
Ran the the Kokkos version with both OpenMP and Cuda (on curie.sandia.gov).

It took some fiddling to make the same kokkos code work for both OpenMP
and Cuda, without having to build all code with nvcc for the Cuda case.

Finally reached a state where the "main" (unit_test_kokkos_saxpy.cpp) is built
with gcc, and sees/holds Kokkos::View objects, but separate functions in other
compilation scopes are responsible for creating the Kokkos::Views and for
performing the saxpy with them.

In the case of using OpenMP, gcc can compile all files. In the Cuda case, nvcc
must compile all files which contain kernels that run on the GPU, and all files
which contain code that passes anything directly to or from the GPU, or launches
any GPU kernels.

None of this code has any effect in Sierra, it must all be built using CMAKE,
and none of the files are referenced in any Sierra Jamfile.

(From Sierra)