MueLu: skips a test if there's no golden output
If the .res file isn't present, the test is skipped.  This is an easy
way to test just Epetra or Tpetra, but not both.  This is exactly what I
wanted to do with additive Schwarz smoothing, as the input deck is
different depending on whether E/Tpetra is being used.

Thanks to Andrey for the suggestion.

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => passed: skipped configure, build, test due to no enabled packages (0.01 min)
1) SERIAL_RELEASE => passed: skipped configure, build, test due to no enabled packages (0.01 min)
2) MPI_SS_DEBUG => FAILED: passed=49,notpassed=2 => Not ready to push! (14.46 min)
3) SERIAL_SS_DEBUG => FAILED: passed=90,notpassed=2 => Not ready to push! (15.36 min)
Other local commits for this build/test group: 0177bc6, 107d2c1, 96845a2
WARNING: Forced the push!