PyTrilinos: Updated build system for release 11.12
New version requirements for Python (2.5) and SWIG (3.0.0).  New build
option PyTrilinos_DOCSTRINGS, which controls whether doxygen is used
to build the Python docstrings for PyTrilinos.  The default is set to
whether doxygen is found or not.

Removed the Python wrappers for Percept.  Percept has been moved out
of STK, so they were also removed from PyTrilinos.  The may return if
Percept is moved directly into Trilinos.