Prevent clang pragmas from being run when building with Intel
Some of the clang pragmas that would treat a header as a system header were
being run when being compiled with the Intel compiler. This is causing a number
of issues for Sierra. This applies a patch from Mike Glass to also check that we
aren't compiling with Intel before setting the pragma.