Update Sierra flat headers list
Removed the Tpetra TSQR directories from the list of files to put in the flat
headers dir for Sierra. The TSQR implementation was moved in Tpetra to its own
subpackage and it shouldn't be needed by Sierra so removing it is the best step
instead of pointing to the new location. If it turns out it is needed later it
can always be added back in.