Disallow NNTI_register_memory() and NNTI_register_segments() for NNTI_SEND_SRC, NNTI_RECV_DST and NNTI_RECV_QUEUE buffer ops.
Some transports require additional memory for send/receive operations.
Since this information is not available to the application, NNTI_alloc()
is required to allocate and register buffers for these operations.

Send/receive operations are not supported on segmented buffers.

This commit disallows NNTI_register_memory() and
NNTI_register_segments() for send/receive buffers.  The error code
NNTI_EINVAL is returned to the app.