Fix race condition between process_event() and is_wr_complete().
A race condition between process_event() and is_wr_complete() could
cause non-progress threads to wait their entire timeout.  The race
occurs when a non-progress thread tests the work request state while
the progress thread is updating the work request state.  This patch
locks access to the state of the work request for completion
operations.  The current fix locks the entire call to
progress_event().  This should be refined.