Sacado:  Rework Promote for proper ViewFad and Expression specializations.
This reworks the various Promote specializations so that (hopefully) all
of the needed specializations for ViewFad and expression templates are
there.  This is fairly complicated, and requires use of SFINAE to
disable some ambiguous partial specializations (thus I likely broke
compatability with gcc 4.1, but I don't see much of a choice).  I also
added more promote tests involving ViewFad and expressions, as well as a
Kokkos test that mixes Fad and ViewFad in an expression.