Finish adding overrides of assignImpl() to all decorators, composites, etc. (Trilinos #5837)
When you add a new virtual function, you need to add it to all of the various
decorator, composite and other types of implicit delegation classes.

In order to test the assignImpl() override for the classes
DefaultColumnwiseMultiVector.

I don't know how good the coverage is here but it should be getting a little
better.

NOTE: I skipped the DefaultClusteredSpmd* classes for now.  These need much
better testing before I mess with them.