Adjust to the new TriBITS layout (TriBITS #26)
Here I have removed TriBITS from being a Trilinos packages to being an extra
repo.  That is just cleaner all the way around.  Now Trilinos automated
testing will clone TriBITS from github and test it after all of the Trilinos
packages.

Note that the functional part of TriBITS is still being snapshotted to
Trilinos/cmake/tribits/ so that no-one needs to clone anything.

Now that TriBITS is being managed as an extra repo for Trilinos testing, we
need to add it to the standard ignores list.