Allow a custom allocator to be injected into the stk-mesh field-data-manager.
No new dependencies are added to stk-mesh.
The field-data-manager class, which is an optional argument to stk-mesh BulkData,
can now take an argument which is an AllocatorAdaptor. This is just a mechanism
for bringing a user-specified allocator type in.

An adagio parser handler is added to allow experimenting with this. The syntax
is:
  developer command: stk field data allocator user simd aligned

This is not on by default, default behavior remains unchanged.

(From Sierra)