Tpetra::MultiVector::putScalar: Now respects DualView last modified.
This commit only affects the Kokkos refactor version of Tpetra.

In order to support hardware with two memory spaces, one of which is
larger (but possibly slower) than the other, MultiVector needs to
implement DualView semantics more carefully.  In particular,
MultiVector methods that modify data should prefer only to touch data
in the last modified memory space.  This commit fixes putScalar so
that it does just that.