Tpetra::MultiVector: Deprecated createMultiVectorFromView.
This nonmember constructor creates a MultiVector that views the data
owned by a Teuchos::ArrayRCP.  This doesn't work with the Kokkos
refactor version of Tpetra, so I'm deprecating it.  The Kokkos refactor
version of MultiVector has constructors that can view a
Kokkos::DualView, which provides equivalent functionality.  I also
commented out the unit tests that exercise this function and the
corresponding function (createVectorFromView) for Vector.