TpetraKernels: Fixed 32-bit issue in MV_Dot() and V_Dot().
MV_Dot() and V_Dot() had an optional argument "n", which is the local
number of rows in the (multi)vector.  If n == -1, the functions use the
Kokkos::View to get the number of rows.  The problem is that n was
declared as int.  This prevents correct use of these functions when the
(multi)vector has 2^{31} or more rows.  This commit makes the optional
argument no longer optional, but changes its type to the size_type
typedef from the Kokkos::View.  This fixes the 32-bit issue.