Ifpack2::Details::Chebyshev: Fix for Scalar=std::complex<T>.
Thanks to Andrey Prokopenko for pointing this out.  Please note that
Ifpack2's Chebyshev implementation only works for real, symmetric
positive definite matrices.  It might _build_ for
Scalar=std::complex<T>, but the algorithm doesn't make sense for that
case.  The right thing to do would be what Belos does, which is provide
a partial specialization for ScalarType = std::complex<T> with a stub
implementation (that builds, but whose constructor throws).  This commit
only provides a "hack" fix that makes the build work.