Ifpack2:  Fix ETI logic for Kokkos wrapper nodes.
Now that TeuchosKokkosCompat is a required dependency for TpetraCore,
the variable TpetraCore_ENABLE_TeuchosKokkosCompat no longer exits which
was causing the instantation files for the wrapper nodes to not be
created.  I just removed this check completely.  I also added a proper
check for the serial device now that there is a cmake variable for it.