Ifpack2: (temporarily) fixed issue with Kokkos::complex* to std::complex* conversion
In a function a ptr to the block of a blockCrs matrix is taken. If the block CrsMatrix
was instantiated with std::complex that getRawPtr thing actually will return Kokkos::complex*
but the matrix_type::scalar_type is std::complex. So I inserted a ptr conversion
to the matrix_type::scalar_type* in front of the getRawPtr function. Probably we should revisit
this fix and fix it inside the block matrix at some point. Mark?