Kokkos: Complex added complex<float> specialisations for atomic_add/assign
Note: the atomic_assign stuff doesn't actually work correct and neither
did it for complex<double> the problem is that its implemented as a
component wise assign which doesn't guarantee that both components come
from the same assignment. We need to add protections in Tpetra
to only allow instantiation of complex for serialwrappernode for now.