Make GlobalMPISession destructor call Kokkos::finalize_all()
Now that the default Tpetra node uses a Kokkos node that requires
Kokkos::Array, you need to call Kokkos::finalize_all() to make sure that all
objects are finalized correctly.

I put in a call to the Kokkos::finalize_all() function in the destructor to
Teuchos::GlobalMPISession.  This avoids warnings about Kokkos objects not
being deallocated correctly.

To make this work, I had to add an optional dependency of TeuchosCore on
KokkosCore.  Optional dependencies are no problem.

I also create the configured file TeuchosCore_config.h and
TeuchosCore_ConfigDefs.hpp.  For Trilinos 12.0, I want to get rid of the
single Teuchos_config.h file and replace it with config files for each
subpackage.