Tpetra: This changes the behaviour of what gets instantiated to reduce memory pressure DownStream
After switching to use Tpetras Macros for instantiation of MueLu, Ifpack2 and Amesos2 users
observed enormous memory requirements for compilation. Reports came if of usage of more than
8GB per MueLu test at link time.

To mitigate this this commit reduces the number of instantiations users will get from Tpetra.
With the exception of Scalar=double and Ordinal=int which are always on unless users explicitly
disable them a user will get something only when explicitly asking for it.

The Options are for Global Ordinals are:
Tpetra_INST_INT_INT (int)
Tpetra_INST_INT_UNSIGNED (unsigned)
Tpetra_INST_INT_LONG (long)
Tpetra_INST_INT_LONG_LONG (long long)

The Options are for Scalars are:
Tpetra_INST_DOUBLE (double)
Tpetra_INST_FLOAT (float)
Tpetra_INST_COMPLEX_FLOAT (std::complex<float>)
Tpetra_INST_COMPLEX_DOUBLE (std::complex<double>)

The Options for nodes are (they only have an effect if Kokkos_Refactor is enabled):
Tpetra_INST_SERIAL (SerialNode (KokkosClassic))
Tpetra_INST_SERIAL (KokkosSerialWrapperNode)
Tpetra_INST_CUDA (KokkosCudaWrapperNode)
Tpetra_INST_OPENMP (KokkosOpenMPWrapperNode)
Tpetra_INST_PTHREAD (KokkosPthreadWrapperNode)

If a user does not set any of the Tpetra_INST options for nodes the behaviour is as follows:
If CUDA is on the node KokkosCudaWrapperNode will be enabled if not
it checks whether OpenMP is on and enables KokkosOpenMPWrapperNode if not
it checks whether Kokkos_ENABLE_Serial is on (which it is by default) and enables
KokkosSerialWrapperNode if not it checks whether pthreads is on and enables that if not it
checks for KokkosClassic SerialNode.

For now pthreads is ranked lower than serial because pthreads in Trilinos is by default on.
So its not a contious choice of the user to enable it.