Tpetra: RTI examples no longer attempt to modify the list of ETI types.
The RTI examples in Tpetra require that both float and double are in the
list of Scalar types over which Tpetra does explicit template
instantiation (ETI).  As a result, the CMake code there used to add
those two types to ${PACKAGE_NAME}_ETI_SCALARS.  This could be
confusing.  This commit changes the CMake code so that it only tests if
those two types are in ${PACKAGE_NAME}_ETI_SCALARS; it no longer
attempts to change that variable.