Tpetra: Fixing the failing thyra test issue caused by changed behaviour
With the new node types in transpose mode y was not overwritten in
y=beta*y+alpha*A*x when beta=0. Because we put no specialisation in for the kernel
with beta=0, y was simply multiplied by 0. The thyra test gives Tpetra a y vector
full of nans though. And that propagates.