Anasazi: Added more output to randomly failing Epetra BKS test.
The Epetra version of a Block Krylov-Schur test randomly fails, by an
amount much more than the tolerance.  I'm not sure why this should be
random, especially since I am not building with OpenMP (the most likely
source of nondeterminism in Epetra).  I added some output to the test to
help us diagnose the issue.