Zoltan2: fixed an incorrect usage scenario of ArrayRCP and ArrayView
Zoltan2 was using a rather complicated way to create a subview of a Tpetra
Mutlivector by first asking for ArrayRCPs of the data when getting
ArrayViews of those, which were put into an array of ArrayViews and then
used to create a new MultiVector. This is not actually valid, since technically
an ArrayRCP reference has to stick around if you create a View from it.
Before switching to the new Kokkos this was implicitly the case since
the Tpetra Multivector had an ArrayRCP to the data. This is not longer the case.
While it is valid to ask for an ArrayRCP (and it is actually still valid even if the
Multivector went away, since it got an deallocator object which keeps the
Kokkos View) that ArrayRCP now has to stick around if you get an ArrayView of it.

Long story short: the usage case in Zoltan2 (getting a subview of a MultiVector
with a new MAP) is actually directly supported by Tpetra. A Multivector has a
function called offsetView which takes a Map and an offset. With offset=0 this
is what the code in Zoltan2 was doing. I commented out the old code and put
the new one in which is valid both with the old node types and the new.
Now all Zoltan2 tests pass.