Updates to cmake files to build PyTrilinos.
PyTrilinos was requiring the python libraries when it didn't really use them so
I disabled the check for them.

PyTrilinos currently depends on the long long instantiations from Tpetra when
ETI is enabled. There is an issue if you enable the Tpetra instantiations but
not the same in Teuchos that Tpetra won't build. Added a configure time tese to
catch this condition and error out with some tips on how to resolve the
situation.