Kokkos: adding malloc<MemorySpace>, free<MemorySpace>, realloc<MemorySpace>
The functions are templated versions of the C-style malloc/free/realloc.
They provide memory coming from one of the memory spaces. Allignments guarantees
are the same as for Kokkos Views. The memory gets tracked, so forgetting to free it
or trying to free it twice will five the appropriate errors.

For malloc there is another overload of the functions which takes a label
void* malloc<class MemorySpace>(const std::string label, size_t size);

Note these functions are only available in C++11 since I wanted to add a default
template argument, which is the memory_space of the DefaultExecutionSpace. That means
int* data = malloc<>(100);
is a valid call.

Furthermore you can use execution spaces as template argumenters (e.g. Kokkos::OpenMP).
It will internally use the memory_space typedef in the template argument.